rename: base -> common 패키지 rename [AIR-144]
